+++ src/components/extra-films.js
@@ -1,0 +1,12 @@
+export // extra films
+  const createExtraFilmTemplate = (title) => {
+    return (
+      `<section class="films-list--extra">
+      <h2 class="films-list__title">${title}</h2>
+
+      <div class="films-list__container">
+      </div>
+    </section>`
+    );
+  };
+
