+++ src/components/films.js
@@ -1,0 +1,8 @@
+export // раздел с фильмами
+  const createAllFilmsTemplate = () => {
+    return (
+      `<section class="films">
+    </section>`
+    );
+  };
+
